refactor common code, change handling of env var
