+++ igniter/tools.py
@@ -248,3 +248,15 @@ def get_openpype_path_from_db(url: str) -> Union[str, None]:
         if os.path.exists(path):
             return path
     return None
+
+
+def load_stylesheet() -> str:
+    """Load css style sheet.
+
+    Returns:
+        str: content of the stylesheet
+
+    """
+    stylesheet_path = Path(__file__).parent.resolve() / "stylesheet.css"
+
+    return stylesheet_path.read_text()
